+++ src/main/java/com/redditapp/redditdownloader/RedditVideoInfo.java
@@ -1,12 +1,12 @@
 package com.redditapp.redditdownloader;
 
 public record RedditVideoInfo(
-        int bitrate,
+        double bitrate,
         String fallbackURL,
         boolean hasAudio,
-        int height,
-        int width,
-        int duration,
+        double height,
+        double width,
+        double duration,
         String dashURL,
         String hlsURL,
         boolean isGif
